1. Converted Nbody routines to use standard (FLOAT) precision rather than double precision to allow single or double precision compilation (still some small problem with sf and su output though) 2. Changed accrete sinks routine to use the tree for quicker evaluation of particles to accrete plus correctly parallelised with OpenMP.

+++ src/GradhSph/GradhSphTree.cpp
@@ -975,7 +975,8 @@ void GradhSphTree<ndim,ParticleType,TreeCell>::UpdateAllSphGravForces
 
       // Compute average/maximum term for computing gravity MAC
       if (gravity_mac == "eigenmac") {
-        for (j=0; j<Nactive; j++) macfactor = max(macfactor,pow(1.0/activepart[j].gpot,twothirds));
+        for (j=0; j<Nactive; j++)
+          macfactor = max(macfactor, pow((FLOAT) 1.0/activepart[j].gpot, twothirds));
       }
 
       // Zero/initialise all summation variables for active particles
